+++ packages/core/src/odata-common/selectable/complex-type-field.ts
@@ -24,8 +24,9 @@ import type { ConstructorOrField } from './constructor-or-field';
  */
 export abstract class ComplexTypeField<
   EntityT extends Entity,
-  ComplexT = any
-> extends Field<EntityT> {
+  ComplexT extends Record<string, any> = any,
+  NullableT extends boolean = false
+> extends Field<EntityT, NullableT> {
   /**
    * @hidden
    * Note that this property is crucial, although not really used.
@@ -39,16 +40,29 @@ export abstract class ComplexTypeField<
   readonly _complexType: ComplexTypeNamespace<ComplexT>;
 
   /**
+   * @deprecated Since v1.27.0. Use other constructors instead.
+   * Creates an instance of ComplexTypeField.
+   *
+   * @param fieldName - Actual name of the field as used in the OData request.
+   * @param fieldOf - Either the parent entity constructor of the parent complex type this field belongs to.
+   */
+  constructor(
+    fieldName: string,
+    fieldOf: ConstructorOrField<EntityT, ComplexT>
+  );
+  /**
    * Creates an instance of ComplexTypeField.
    *
    * @param fieldName - Actual name of the field as used in the OData request.
    * @param fieldOf - Either the parent entity constructor of the parent complex type this field belongs to.
    * @param complexType - The complex type of the complex type property represented by this.
+   * @param isNullable - Whether the field can have the value `null`.
    */
   constructor(
     fieldName: string,
     fieldOf: ConstructorOrField<EntityT, ComplexT>,
-    complexType?: ComplexTypeNamespace<ComplexT>
+    complexType: ComplexTypeNamespace<ComplexT>,
+    isNullable?: NullableT
   );
 
   /**
@@ -72,13 +86,15 @@ export abstract class ComplexTypeField<
    * @param fieldName - Actual name of the field as used in the OData request.
    * @param fieldOf - Either the parent entity constructor of the parent complex type this field belongs to.
    * @param complexTypeOrName - The complex type of the complex type property represented by this or the name of the type of the field according to the metadata description. Using the name here is deprecated.
+   * @param isNullable - Whether the field can have the value `null`.
    */
   constructor(
     fieldName: string,
     readonly fieldOf: ConstructorOrField<EntityT, ComplexT>,
-    complexTypeOrName?: ComplexTypeNamespace<ComplexT> | string
+    complexTypeOrName?: ComplexTypeNamespace<ComplexT> | string,
+    isNullable: NullableT = false as NullableT
   ) {
-    super(fieldName, getEntityConstructor(fieldOf));
+    super(fieldName, getEntityConstructor(fieldOf), isNullable);
     if (typeof complexTypeOrName === 'string') {
       this.complexTypeName = complexTypeOrName;
     } else if (isComplexTypeNameSpace(complexTypeOrName)) {