+++ packages/core/src/odata-common/selectable/constructor-or-field.ts
@@ -1,5 +1,5 @@
 import { Constructable, Entity } from '../entity';
-import { ComplexTypeField } from './complex-type-field';
+import type { ComplexTypeField } from './complex-type-field';
 
 /**
  * Union type to represent the parent of a field. This can either be an entity constructor or a complex type field.