+++ packages/core/src/odata-common/selectable/custom-field.ts
@@ -1,120 +1,134 @@
 import { Entity, Constructable } from '../entity';
-import { BigNumberField } from './big-number-field';
-import { BooleanField } from './boolean-field';
-import { DateField } from './date-field';
+import { EdmTypeField } from './edm-type-field';
 import { Field } from './field';
-import { NumberField } from './number-field';
-import { StringField } from './string-field';
 
-export class CustomField<EntityT extends Entity> extends Field<EntityT> {
+export class CustomField<
+  EntityT extends Entity,
+  NullableT extends boolean = false
+> extends Field<EntityT, NullableT> {
   constructor(
-    readonly _fieldName: string,
-    readonly _entityConstructor: Constructable<EntityT>
+    fieldName: string,
+    entityConstructor: Constructable<EntityT>,
+    isNullable: NullableT = false as NullableT
   ) {
-    super(_fieldName, _entityConstructor);
+    super(fieldName, entityConstructor, isNullable);
   }
 
-  edmString(): StringField<EntityT> {
-    return new StringField(
+  edmString(): EdmTypeField<EntityT, 'Edm.String', NullableT> {
+    return new EdmTypeField(
       this._fieldName,
       this._entityConstructor,
-      'Edm.String'
+      'Edm.String',
+      this._isNullable
     );
   }
 
-  edmBoolean(): BooleanField<EntityT> {
-    return new BooleanField(
+  edmBoolean(): EdmTypeField<EntityT, 'Edm.Boolean', NullableT> {
+    return new EdmTypeField(
       this._fieldName,
       this._entityConstructor,
-      'Edm.Boolean'
+      'Edm.Boolean',
+      this._isNullable
     );
   }
 
-  edmGuid(): StringField<EntityT> {
-    return new StringField(
+  edmGuid(): EdmTypeField<EntityT, 'Edm.Guid', NullableT> {
+    return new EdmTypeField(
       this._fieldName,
       this._entityConstructor,
-      'Edm.Guid'
+      'Edm.Guid',
+      this._isNullable
     );
   }
 
-  edmDecimal(): BigNumberField<EntityT> {
-    return new BigNumberField(
+  edmDecimal(): EdmTypeField<EntityT, 'Edm.Decimal', NullableT> {
+    return new EdmTypeField(
       this._fieldName,
       this._entityConstructor,
-      'Edm.Decimal'
+      'Edm.Decimal',
+      this._isNullable
     );
   }
 
-  edmInt16(): NumberField<EntityT> {
-    return new NumberField(
+  // TODO: GL
+  edmInt16(): EdmTypeField<EntityT, 'Edm.Int16', NullableT> {
+    return new EdmTypeField(
       this._fieldName,
       this._entityConstructor,
-      'Edm.Int16'
+      'Edm.Int16',
+      this._isNullable
     );
   }
 
-  edmInt32(): NumberField<EntityT> {
-    return new NumberField(
+  edmInt32(): EdmTypeField<EntityT, 'Edm.Int32', NullableT> {
+    return new EdmTypeField(
       this._fieldName,
       this._entityConstructor,
-      'Edm.Int32'
+      'Edm.Int32',
+      this._isNullable
     );
   }
 
-  edmInt64(): BigNumberField<EntityT> {
-    return new BigNumberField(
+  edmInt64(): EdmTypeField<EntityT, 'Edm.Int64', NullableT> {
+    return new EdmTypeField(
       this._fieldName,
       this._entityConstructor,
-      'Edm.Int64'
+      'Edm.Int64',
+      this._isNullable
     );
   }
 
-  edmSingle(): NumberField<EntityT> {
-    return new NumberField(
+  edmSingle(): EdmTypeField<EntityT, 'Edm.Single', NullableT> {
+    return new EdmTypeField(
       this._fieldName,
       this._entityConstructor,
-      'Edm.Single'
+      'Edm.Single',
+      this._isNullable
     );
   }
 
-  edmDouble(): NumberField<EntityT> {
-    return new NumberField(
+  edmDouble(): EdmTypeField<EntityT, 'Edm.Double', NullableT> {
+    return new EdmTypeField(
       this._fieldName,
       this._entityConstructor,
-      'Edm.Double'
+      'Edm.Double',
+      this._isNullable
     );
   }
 
-  edmByte(): NumberField<EntityT> {
-    return new NumberField(
+  edmByte(): EdmTypeField<EntityT, 'Edm.Byte', NullableT> {
+    return new EdmTypeField(
       this._fieldName,
       this._entityConstructor,
-      'Edm.Byte'
+      'Edm.Byte',
+      this._isNullable
     );
   }
 
-  edmSByte(): NumberField<EntityT> {
-    return new NumberField(
+  edmSByte(): EdmTypeField<EntityT, 'Edm.SByte', NullableT> {
+    return new EdmTypeField(
       this._fieldName,
       this._entityConstructor,
-      'Edm.SByte'
+      'Edm.SByte',
+      this._isNullable
     );
   }
 
-  edmDateTimeOffset(): DateField<EntityT> {
-    return new DateField(
+  edmDateTimeOffset(): EdmTypeField<EntityT, 'Edm.DateTimeOffset', NullableT> {
+    return new EdmTypeField(
       this._fieldName,
       this._entityConstructor,
-      'Edm.DateTimeOffset'
+      'Edm.DateTimeOffset',
+      this._isNullable
     );
   }
 
-  edmBinary(): StringField<EntityT> {
-    return new StringField(
+  edmBinary(): EdmTypeField<EntityT, 'Edm.Binary', NullableT> {
+    return new EdmTypeField(
       this._fieldName,
       this._entityConstructor,
-      'Edm.Binary'
+      'Edm.Binary',
+      this._isNullable
     );
   }
 }