+++ packages/core/src/odata-common/selectable/greater-or-less.ts
@@ -1,0 +1,59 @@
+/* eslint-disable max-classes-per-file */
+import { Filter } from '../filter';
+import { Entity } from '../entity';
+import { EdmTypeShared } from '../edm-types';
+import { EdmTypeField, FieldTypeByEdmType } from './edm-type-field';
+
+export class GreaterOrLessEdmTypeField<
+  EntityT extends Entity,
+  EdmT extends EdmTypeShared<'any'>,
+  NullableT extends boolean = false
+> extends EdmTypeField<EntityT, EdmT, NullableT> {
+  /**
+   * Creates an instance of Filter for this field and the given value using the operator 'gt', i.e. `>`.
+   *
+   * @param value - Value to be used in the filter
+   * @returns The resulting filter
+   */
+  greaterThan(
+    value: FieldTypeByEdmType<EdmT, NullableT>
+  ): Filter<EntityT, FieldTypeByEdmType<EdmT, NullableT>> {
+    return new Filter(this.fieldPath(), 'gt', value, this.edmType);
+  }
+
+  /**
+   * Creates an instance of Filter for this field and the given value using the operator 'ge', i.e. `>=`.
+   *
+   * @param value - Value to be used in the filter
+   * @returns The resulting filter
+   */
+  greaterOrEqual(
+    value: FieldTypeByEdmType<EdmT, NullableT>
+  ): Filter<EntityT, FieldTypeByEdmType<EdmT, NullableT>> {
+    return new Filter(this.fieldPath(), 'ge', value, this.edmType);
+  }
+
+  /**
+   * Creates an instance of Filter for this field and the given value using the operator 'lt', i.e. `<`.
+   *
+   * @param value - Value to be used in the filter
+   * @returns The resulting filter
+   */
+  lessThan(
+    value: FieldTypeByEdmType<EdmT, NullableT>
+  ): Filter<EntityT, FieldTypeByEdmType<EdmT, NullableT>> {
+    return new Filter(this.fieldPath(), 'lt', value, this.edmType);
+  }
+
+  /**
+   * Creates an instance of Filter for this field and the given value using the operator 'le', i.e. `<=`.
+   *
+   * @param value - Value to be used in the filter
+   * @returns The resulting filter
+   */
+  lessOrEqual(
+    value: FieldTypeByEdmType<EdmT, NullableT>
+  ): Filter<EntityT, FieldTypeByEdmType<EdmT, NullableT>> {
+    return new Filter(this.fieldPath(), 'le', value, this.edmType);
+  }
+}