+++ packages/core/src/odata-v2/entity.ts
@@ -5,11 +5,15 @@ import { CustomField } from './selectable';
  * Super class for all representations of OData v2 entity types.
  */
 export class Entity extends EntityBase {
-  protected static customFieldSelector<EntityT extends Entity>(
+  protected static customFieldSelector<
+    EntityT extends Entity,
+    NullableT extends boolean = false
+  >(
     fieldName: string,
-    entityConstructor: Constructable<EntityT>
-  ): CustomField<EntityT> {
-    return new CustomField(fieldName, entityConstructor);
+    entityConstructor: Constructable<EntityT>,
+    isNullable: NullableT = false as NullableT
+  ): CustomField<EntityT, NullableT> {
+    return new CustomField(fieldName, entityConstructor, isNullable);
   }
 
   readonly _oDataVersion: 'v2' = 'v2';