+++ packages/core/src/odata-v2/selectable/custom-field.ts
@@ -1,25 +1,27 @@
 import {
   CustomField as CustomFieldBase,
-  DateField,
-  TimeField
+  EdmTypeField
 } from '../../odata-common';
 import type { Entity } from '../entity';
 
 export class CustomField<
-  EntityT extends Entity
-> extends CustomFieldBase<EntityT> {
-  edmDateTime(): DateField<EntityT> {
-    return new DateField<EntityT>(
+  EntityT extends Entity,
+  NullableT extends boolean = false
+> extends CustomFieldBase<EntityT, NullableT> {
+  edmDateTime(): EdmTypeField<EntityT, 'Edm.DateTime', NullableT> {
+    return new EdmTypeField(
       this._fieldName,
       this._entityConstructor,
-      'Edm.DateTime'
+      'Edm.DateTime',
+      this._isNullable
     );
   }
-  edmTime(): TimeField<EntityT> {
-    return new TimeField<EntityT>(
+  edmTime(): EdmTypeField<EntityT, 'Edm.Time', NullableT> {
+    return new EdmTypeField(
       this._fieldName,
       this._entityConstructor,
-      'Edm.Time'
+      'Edm.Time',
+      this._isNullable
     );
   }
 }